Initial worker for reporters and move executors to its own package.
